reorganized instruments files
